requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
CacheControl==0.14.2
cachetools==5.5.2
cbor2==5.6.5
certifi==2025.1.31
certvalidator==0.11.1
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
cryptography==44.0.2
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-appconf==1.1.0
django-cors-headers==4.7.0
django-filter==25.1
django-imagekit==5.0.0
django-ipware==7.0.1
django-push-notifications==3.2.1
django-rest-knox==5.0.2
djangorestframework==3.16.0
djangorestframework-recaptcha==0.2.0
drf-spectacular==0.28.0
fcm-django==2.2.1
firebase-admin==6.7.0
frozenlist==1.6.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
msgpack==1.1.0
multidict==6.4.3
oauthlib==3.2.2
oscrypto==1.3.0
pilkit==3.0
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
py-webauthn==0.0.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-decouple==3.8
python-ipware==3.0.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9
sqlparse==0.5.3
stripe==12.0.0
swapper==1.4.0
twilio==9.6.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
webauthn==2.5.2
yarl==1.20.0
django-storages==1.14.6
boto3==1.41.5